feat: add provider SVG icons and color-coded tooltip bars
Replace text labels (Claude/Codex) with OpenAI and Claude brand SVG
icons in both the status bar and tooltip header. Add green/yellow/red
color coding to tooltip progress bars based on remaining capacity.
Extract tooltip components to a separate file to stay under max-lines.
